+++ src/main/java/net/oneandone/lavender/config/Target.java
@@ -22,12 +22,10 @@ import java.io.IOException;
 public class Target {
     public final Cluster cluster;
     public final Docroot docroot;
-    public final Alias alias;
 
-    public Target(Cluster cluster, Docroot docroot, Alias alias) {
+    public Target(Cluster cluster, Docroot docroot) {
         this.cluster = cluster;
         this.docroot = docroot;
-        this.alias = alias;
     }
 
     public Distributor open(Pool pool, String indexName) throws IOException {